Implementing the Destroy Operation of ProductCategory

+++ app/Http/Controllers/Product/ProductCategoryController.php
@@ -55,6 +55,10 @@ class ProductCategoryController extends ApiController
      */
     public function destroy(Product $product, Category $category)
     {
+        if (!$product->categories()->find($category->id)){
+            return $this->errorResponse("The specific category is not a category of this product", 404);
+        }
+
         $product->categories()->detach($category->id);
 
         return $this->showOne($product);
